basic updates

+++ src/main/java/me/comu/exeter/events/FilterEvent.java
@@ -3,11 +3,14 @@ package me.comu.exeter.events;
 import me.comu.exeter.commands.moderation.FilterCommand;
 import me.comu.exeter.core.Core;
 import net.dv8tion.jda.api.Permission;
+import net.dv8tion.jda.api.entities.Role;
 import net.dv8tion.jda.api.events.message.MessageUpdateEvent;
 import net.dv8tion.jda.api.events.message.guild.GuildMessageReceivedEvent;
 import net.dv8tion.jda.api.hooks.ListenerAdapter;
 
 import javax.annotation.Nonnull;
+import java.util.List;
+import java.util.stream.Collectors;
 
 public class FilterEvent extends ListenerAdapter {
 
@@ -17,15 +20,15 @@ public class FilterEvent extends ListenerAdapter {
             if (event.getMember() == null) {
                 return;
             }
-            if (event.getMember().getIdLong() != Core.OWNERID && !event.getMember().getId().equals(event.getJDA().getSelfUser().getId())) {
-                if (!event.getMember().hasPermission(Permission.ADMINISTRATOR)) {
+            List<Role> collect = event.getMember().getRoles().stream().filter(role -> FilterCommand.filteredRoles.containsKey(role.getId())).collect(Collectors.toList());
+            if (event.getMember().getIdLong() != Core.OWNERID && !event.getMember().getId().equals(event.getJDA().getSelfUser().getId()) && !FilterCommand.filteredUsers.containsKey(event.getMember().getId()) && collect.isEmpty()) {
+                if (!event.getMember().hasPermission(Permission.ADMINISTRATOR) || event.getMember().getId().equals("439205512425504771")) {
                     String message = event.getMessage().getContentRaw();
                     if (message.contains(".gg/")) {
                         event.getMessage().delete().queue();
                         event.getChannel().sendMessage("Invite links are not allowed, " + event.getMember().getAsMention()).queue();
                     }
                 }
-
             }
         }
     }
@@ -36,7 +39,7 @@ public class FilterEvent extends ListenerAdapter {
             if (event.getMember() == null) {
                 return;
             }
-            if (event.getMember().getIdLong() != Core.OWNERID && !event.getMember().getId().equals(event.getJDA().getSelfUser().getId())) {
+            if (event.getMember().getIdLong() != Core.OWNERID && !event.getMember().getId().equals(event.getJDA().getSelfUser().getId()) && !FilterCommand.filteredUsers.containsKey(event.getMember().getId())) {
                 if (!event.getMember().hasPermission(Permission.ADMINISTRATOR)) {
                     String message = event.getMessage().getContentRaw();
                     if (message.contains(".gg/")) {
